tenant-onboarding-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic[email]
google-cloud-firestore==2.13.1
google-cloud-storage==2.13.0
google-cloud-bigquery==3.13.0
google-cloud-run==0.10.0
google-cloud-build==3.22.0
google-cloud-secret-manager==2.16.4
google-cloud-logging==3.8.0
structlog==23.1.0
httpx==0.25.2
python-multipart==0.0.6
